app/requirements.txt
pandas
numpy
langchain-community
sentence-transformers
langchain_experimental
transformers
chromadb
tavily-python
torch
datasets
tiktoken
groq
python-dotenv
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
psycopg2-binary # PostgreSQL driver 
feedparser
requests
Beautifulsoup4
html5lib
apscheduler # For background tasks
aiohttp # For async HTTP requests